move exception handling to network layer
