+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/network/credentialOperations/FetchPresentationRequestNetworkOperation.kt
@@ -11,7 +11,9 @@ import com.microsoft.did.sdk.crypto.protocols.jose.jws.JwsToken
 import com.microsoft.did.sdk.datasource.network.GetNetworkOperation
 import com.microsoft.did.sdk.datasource.network.apis.ApiProvider
 import com.microsoft.did.sdk.util.controlflow.DidInHeaderAndPayloadNotMatching
+import com.microsoft.did.sdk.util.controlflow.ExpiredTokenException
 import com.microsoft.did.sdk.util.controlflow.InvalidSignatureException
+import com.microsoft.did.sdk.util.controlflow.NotFoundException
 import com.microsoft.did.sdk.util.controlflow.PresentationException
 import com.microsoft.did.sdk.util.controlflow.Result
 import kotlinx.serialization.json.Json
@@ -29,6 +31,24 @@ class FetchPresentationRequestNetworkOperation(
     override suspend fun onSuccess(response: Response<String>): Result<PresentationRequestContent> {
         val jwsTokenString = response.body() ?: throw PresentationException("No Presentation Request in Body.")
         return verifyAndUnwrapPresentationRequest(jwsTokenString)
+    }
+
+    override fun onFailure(response: Response<String>): Result<Nothing> {
+        val result = super.onFailure(response)
+        when (val sdkException = (result as Result.Failure).payload) {
+            is NotFoundException -> {
+                val exception = ExpiredTokenException("The request has expired.", false)
+                exception.apply {
+                    correlationVector = sdkException.correlationVector
+                    errorBody = sdkException.errorBody
+                    errorCode = sdkException.errorCode
+                    innerErrorCodes = sdkException.innerErrorCodes
+                    errorMessage = sdkException.errorMessage
+                }
+                return Result.Failure(exception)
+            }
+        }
+        return result
     }
 
     private suspend fun verifyAndUnwrapPresentationRequest(jwsTokenString: String): Result<PresentationRequestContent> {
